``` refactor: reorganize pages-setting API structure by moving about-us specific code to dedicated module
